+++ src/modules/voxelgenerator/TreeGenerator.cpp
@@ -8,13 +8,13 @@ namespace voxelgenerator {
 namespace tree {
 
 Tree::Tree(const glm::ivec3& position, int trunkHeight, int branchLength,
-	int crownWidth, int crownHeight, int crownDepth, float branchSize, int seed, float trunkSizeFactor) :
+	int crownWidth, int crownHeight, int crownDepth, float branchSize, unsigned int seed, float trunkSizeFactor) :
 			SpaceColonization(glm::ivec3(position.x, position.y + trunkHeight, position.z),
 					branchLength, crownWidth, crownHeight, crownDepth, branchSize, seed),
 			_trunkHeight(trunkHeight), _trunkSizeFactor(trunkSizeFactor) {
-	_root->_position.y -= trunkHeight;
-	_position.y -= trunkHeight;
-	generateBranches(_branches, glm::up, _trunkHeight, _branchLength);
+	_root->_position.y -= (float)trunkHeight;
+	_position.y -= (float)trunkHeight;
+	generateBranches(_branches, glm::up, (float)_trunkHeight, (float)_branchLength);
 }
 
 // TODO: use the PoolAllocator here
